# docs/swagger.yaml
basePath: /api/v0
definitions:
  agent.ActionConfig:
    properties:
      attachments:
        description: Attachments at action scope
        items: {}
        type: array
      cwd:
        $ref: '#/definitions/core.PathCWD'
      id:
        description: |-
          Unique identifier for the action within the agent's scope.
          Used to invoke specific actions programmatically.

          - **Examples:** `"analyze-code"`, `"generate-summary"`, `"validate-data"`
        type: string
      input:
        allOf:
          - $ref: '#/definitions/schema.Schema'
        description: |-
          JSON Schema defining the expected input parameters for this action.
          Enables validation and type checking of inputs before execution.

          If `nil`, the action accepts any input format without validation.

          **Schema format:** JSON Schema Draft 7
      json_mode:
        description: |-
          Forces JSON-formatted output for this specific action.
          When `true`, the agent must return valid JSON that conforms to the output schema.

          **Note:** If an `OutputSchema` is defined, JSON mode is automatically enabled.

          ⚠️ **Trade-off:** Enabling JSON mode may limit the agent's ability to provide
          explanatory text or reasoning alongside the structured output.
        type: boolean
      output:
        allOf:
          - $ref: '#/definitions/schema.Schema'
        description: |-
          JSON Schema defining the expected output format from this action.
          Used for validating agent responses and ensuring consistent output structure.

          If `nil`, no output validation is performed.

          **Schema format:** JSON Schema Draft 7
      prompt:
        description: |-
          Detailed instructions for the agent when executing this action.
          Should clearly define the expected behavior, output format, and any constraints.

          **Best practices:**
          - Be specific about the desired outcome
          - Include examples if complex formatting is required
          - Define clear success criteria
          - Specify any limitations or boundaries
        type: string
      with:
        allOf:
          - $ref: '#/definitions/core.Input'
        description: |-
          Default parameters to provide to the action.
          These are merged with runtime parameters, with runtime values taking precedence.

          **Use cases:**
          - Setting default configuration options
          - Providing constant context values
          - Pre-filling common parameters
    required:
      - prompt
    type: object
  agent.Config:
    properties:
      actions:
        description: |-
          Structured actions the agent can perform with defined input/output schemas.
          Actions provide type-safe interfaces for specific agent capabilities.

          **Example:**
          ```yaml
          actions:
            - id: "review-code"
              prompt: |
                Analyze code {{.input.code}} for quality and improvements
              json_mode: true
              input:
                type: "object"
                properties:
                  code:
                    type: "string"
                    description: "The code to review"
              output:
                type: "object"
                properties:
                  quality:
                    type: "string"
                    description: "The quality of the code"
          ```

          $ref: inline:#action-configuration
        items:
          $ref: '#/definitions/agent.ActionConfig'
        type: array
      attachments:
        description: Attachments declared at the agent scope.
        items: {}
        type: array
      cwd:
        $ref: '#/definitions/core.PathCWD'
      env:
        allOf:
          - $ref: '#/definitions/core.EnvMap'
        description: |-
          Environment variables available during agent execution.
          Used for configuration, secrets, and runtime settings.

          **Example:**
          ```yaml
          env:
            API_KEY: "{{.env.OPENAI_API_KEY}}"
            DEBUG_MODE: "true"
          ```
      id:
        description: |-
          Unique identifier for the agent within the project scope.
          Used for referencing the agent in workflows and other configurations.

          - **Examples:** `"code-assistant"`, `"data-analyst"`, `"customer-support"`
        type: string
      instructions:
        description: |-
          Provider configuration is now expressed through the polymorphic `Model` field.
          The previous `Config core.ProviderConfig` field has been removed.
          System instructions that define the agent's personality, behavior, and constraints.
          These instructions guide how the agent interprets tasks and generates responses.

          **Best practices:**
          - Be clear and specific about the agent's role
          - Define boundaries and ethical guidelines
          - Include domain-specific knowledge or constraints
          - Use markdown formatting for better structure
        type: string
      json_mode:
        description: |-
          Forces the agent to always respond in valid JSON format.
          When enabled, the agent's responses must be parseable JSON objects.

          **Use cases:**
          - API integrations requiring structured data
          - Automated processing of agent outputs
          - Ensuring consistent response formats

          ⚠️ **Note:** May limit the agent's ability to provide explanatory text
        type: boolean
      max_iterations:
        description: |-
          Maximum number of reasoning iterations the agent can perform.
          The agent may self-correct and refine its response across multiple iterations
          to improve accuracy and address complex multi-step problems.

          **Default:** `5` iterations

          **Trade-offs:**
          - Higher values enable more thorough problem-solving and self-correction
          - Each iteration consumes additional tokens and increases response latency
          - Configure based on task complexity, accuracy requirements, and cost constraints
        type: integer
      mcps:
        description: |-
          Model Context Protocol (MCP) server configurations.
          MCPs provide standardized interfaces for extending agent capabilities
          with external services and data sources through protocol-based communication.

          **Common MCP integrations:**
          - Database connectors (PostgreSQL, Redis, MongoDB)
          - Search engines (Elasticsearch, Solr)
          - Knowledge bases (vector databases, documentation systems)
          - External APIs (REST, GraphQL, gRPC services)

          MCPs support both stdio and HTTP transport protocols.
        items:
          $ref: '#/definitions/mcp.Config'
        type: array
      memory:
        description: |-
          Memory references enabling the agent to access persistent context.
          Memory provides stateful interactions across workflow steps and sessions.

          **Configuration format:**
          ```yaml
          memory:
            - id: "user_context"  # Memory resource ID
              key: "user:{{.user_id}}"  # Dynamic key with template
              mode: "read-write"  # Access mode (default: "read-write")
          ```

          **Access modes:**
          - `"read-write"`: Full access to read and modify memory
          - `"read-only"`: Can only read existing memory entries
        items:
          $ref: '#/definitions/core.MemoryReference'
        type: array
      model:
        allOf:
          - $ref: '#/definitions/agent.Model'
        description: |-
          Model selects which LLM model to use.
          Supports two forms:
            - string: a model ID to be resolved via the ResourceStore (e.g. "openai-gpt-4o-mini")
            - object: an inline core.ProviderConfig with provider/model/params

          During compile/link, string refs are resolved and merged with inline
          fields following project precedence rules. Defaults are filled from the
          project when neither ref nor inline identity is provided.
      resource:
        description: |-
          Resource identifier for the autoloader system (must be `"agent"`).
          This field enables automatic discovery and registration of agent configurations.
        type: string
      tools:
        description: |-
          Tools available to the agent for extending its capabilities.
          When tools are defined, the agent automatically has `toolChoice` set to `"auto"`,
          enabling autonomous tool selection and invocation during task execution.

          **Tool types supported:**
          - File system operations (read, write, list)
          - API integrations (HTTP requests, webhooks)
          - Data processing utilities (parsing, transformation)
          - Custom business logic (TypeScript/JavaScript execution)

          Tools are referenced by ID and can be shared across multiple agents.
        items:
          $ref: '#/definitions/tool.Config'
        type: array
      with:
        allOf:
          - $ref: '#/definitions/core.Input'
        description: |-
          Default input parameters passed to the agent on every invocation.
          These values are merged with runtime inputs, with runtime values taking precedence.

          **Use cases:**
          - Setting default configuration values
          - Providing constant context or settings
          - Injecting workflow-level parameters
    required:
      - id
      - instructions
    type: object
  agent.Model:
    properties:
      config:
        allOf:
          - $ref: '#/definitions/core.ProviderConfig'
        description: |-
          Config holds an inline provider configuration when specified as a mapping.
          When Ref is set, Config may be merged with the resolved model during linking.
    type: object
  agentrouter.AgentDTO:
    properties:
      env:
        additionalProperties: {}
        type: object
      id:
        type: string
      instructions:
        type: string
      model:
        additionalProperties: {}
        type: object
      resource:
        type: string
      with:
        additionalProperties: {}
        type: object
    type: object
  agentrouter.AgentListItem:
    properties:
      env:
        additionalProperties: {}
        type: object
      etag:
        example: abc123
        type: string
      id:
        type: string
      instructions:
        type: string
      model:
        additionalProperties: {}
        type: object
      resource:
        type: string
      with:
        additionalProperties: {}
        type: object
    type: object
  agentrouter.AgentsListResponse:
    properties:
      agents:
        items:
          $ref: '#/definitions/agentrouter.AgentListItem'
        type: array
      page:
        $ref: '#/definitions/router.PageInfoDTO'
    type: object
...
  core.EnvMap:
    additionalProperties:
      type: string
    type: object
...
  core.Input:
    additionalProperties: {}
    type: object
  core.MemoryReference:
    properties:
      id:
        type: string
      key:
        description: |-
          Key is a template string that resolves to the actual memory instance key.
          e.g., "support-{{ .workflow.input.conversationId }}"
        type: string
      mode:
        description: Mode defines access permissions (e.g., "read-write", "read-only").
        enum:
          - read-write
          - read-only
        type: string
    required:
      - id
    type: object
...
  core.PathCWD:
    properties:
      path:
        description: Path holds the absolute working directory.
        type: string
    type: object
...
  router.ErrorInfo:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
...
  router.PageInfoDTO:
    properties:
      limit:
        example: 50
        type: integer
      next_cursor:
        example: v2:after:tool-001
        type: string
      prev_cursor:
        example: v2:before:tool-000
        type: string
      total:
        example: 2
        type: integer
    type: object
  router.ProblemDocument:
    properties:
      code:
        example: invalid_cursor
        type: string
      detail:
        example: Invalid cursor parameter
        type: string
      instance:
        example: /api/v0/workflows
        type: string
      status:
        example: 400
        type: integer
      title:
        example: Bad Request
        type: string
      type:
        example: about:blank
        type: string
    type: object
  router.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/router.ErrorInfo'
      message:
        type: string
      status:
        type: integer
    type: object
...
  schema.Schema:
    additionalProperties: {}
    type: object
...
info:
  contact:
    email: support@compozy.com
    name: Compozy Support
    url: https://github.com/compozy/compozy
  description: Compozy is a Next-level Agentic Orchestration Platform, tasks, and tools. Authentication
    is configurable via the SERVER_AUTH_ENABLED environment variable. When enabled, all endpoints require
    Bearer token authentication. When disabled (typical for development), endpoints are accessible without
    authentication.
  license:
    name: BSL-1.1
    url: https://github.com/compozy/compozy/blob/main/LICENSE
  termsOfService: https://github.com/compozy/compozy
  title: Compozy API
  version: "1.0"
paths:
...
  /agents:
    get:
      consumes:
        - application/json
      description: List agents with cursor pagination. Optionally filter by workflow usage.
      parameters:
        - description: Project override
          example: '"demo"'
          in: query
          name: project
          type: string
        - description: Return only agents referenced by the given workflow
          example: '"wf1"'
          in: query
          name: workflow_id
          type: string
        - description: Page size (max 500)
          example: 50
          in: query
          name: limit
          type: integer
        - description: Opaque pagination cursor
          in: query
          name: cursor
          type: string
        - description: Filter by agent ID prefix
          in: query
          name: q
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Agents retrieved
          headers:
            Link:
              description: RFC 8288 pagination links for next/prev
              type: string
            RateLimit-Limit:
              description: Requests allowed in the current window
              type: string
            RateLimit-Remaining:
              description: Remaining requests in the current window
              type: string
            RateLimit-Reset:
              description: Seconds until the window resets
              type: string
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  data:
                    $ref: '#/definitions/agentrouter.AgentsListResponse'
                type: object
        "400":
          description: Invalid cursor
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "404":
          description: Workflow not found
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.ProblemDocument'
      summary: List agents
      tags:
        - agents
  /agents/{agent_id}:
    delete:
      description: Delete an agent configuration. Returns conflict when referenced.
      parameters:
        - description: Agent ID
          example: '"assistant"'
          in: path
          name: agent_id
          required: true
          type: string
        - description: Project override
          example: '"demo"'
          in: query
          name: project
          type: string
      produces:
        - application/json
      responses:
        "204":
          description: No Content
          headers:
            RateLimit-Limit:
              description: Requests allowed in the current window
              type: string
            RateLimit-Remaining:
              description: Remaining requests in the current window
              type: string
            RateLimit-Reset:
              description: Seconds until the window resets
              type: string
          schema:
            type: string
        "404":
          description: Agent not found
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "409":
          description: Agent referenced
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.ProblemDocument'
      summary: Delete agent
      tags:
        - agents
    get:
      consumes:
        - application/json
      description: Retrieve an agent configuration by ID.
      parameters:
        - description: Agent ID
          example: '"assistant"'
          in: path
          name: agent_id
          required: true
          type: string
        - description: Project override
          example: '"demo"'
          in: query
          name: project
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Agent retrieved
          headers:
            ETag:
              description: Strong entity tag for concurrency control
              type: string
            RateLimit-Limit:
              description: Requests allowed in the current window
              type: string
            RateLimit-Remaining:
              description: Remaining requests in the current window
              type: string
            RateLimit-Reset:
              description: Seconds until the window resets
              type: string
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  data:
                    $ref: '#/definitions/agentrouter.AgentDTO'
                type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "404":
          description: Agent not found
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.ProblemDocument'
      summary: Get agent
      tags:
        - agents
    put:
      consumes:
        - application/json
      description: Create an agent when absent or update an existing one using strong ETag concurrency.
      parameters:
        - description: Agent ID
          example: '"assistant"'
          in: path
          name: agent_id
          required: true
          type: string
        - description: Project override
          example: '"demo"'
          in: query
          name: project
          type: string
        - description: Strong ETag for optimistic concurrency
          example: '"\"abc123\""'
          in: header
          name: If-Match
          type: string
        - description: Agent configuration payload
          in: body
          name: payload
          required: true
          schema:
            additionalProperties: true
            type: object
      produces:
        - application/json
      responses:
        "200":
          description: Agent updated
          headers:
            ETag:
              description: Strong entity tag for concurrency control
              type: string
            RateLimit-Limit:
              description: Requests allowed in the current window
              type: string
            RateLimit-Remaining:
              description: Remaining requests in the current window
              type: string
            RateLimit-Reset:
              description: Seconds until the window resets
              type: string
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  data:
                    $ref: '#/definitions/agentrouter.AgentDTO'
                type: object
        "201":
          description: Agent created
          headers:
            ETag:
              description: Strong entity tag for concurrency control
              type: string
            Location:
              description: Absolute URL for the agent
              type: string
            RateLimit-Limit:
              description: Requests allowed in the current window
              type: string
            RateLimit-Remaining:
              description: Remaining requests in the current window
              type: string
            RateLimit-Reset:
              description: Seconds until the window resets
              type: string
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  data:
                    $ref: '#/definitions/agentrouter.AgentDTO'
                type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "404":
          description: Agent not found
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "409":
          description: Agent referenced
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "412":
          description: ETag mismatch
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.ProblemDocument'
      summary: Create or update agent
      tags:
        - agents
...
  /workflows/{workflow_id}/agents:
    get:
      consumes:
        - application/json
      description: Retrieve a list of all available agent configurations
      parameters:
        - description: Workflow ID
          in: path
          name: workflow_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Agents retrieved successfully
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  data:
                    properties:
                      agents:
                        items:
                          $ref: '#/definitions/agent.Config'
                        type: array
                    type: object
                type: object
        "500":
          description: Internal server error
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  error:
                    $ref: '#/definitions/router.ErrorInfo'
                type: object
      summary: List all agents
      tags:
        - agents
  /workflows/{workflow_id}/agents/{agent_id}:
    get:
      consumes:
        - application/json
      description: Retrieve a specific agent configuration by its ID
      parameters:
        - description: Workflow ID
          in: path
          name: workflow_id
          required: true
          type: string
        - description: Agent ID
          example: '"code-assistant"'
          in: path
          name: agent_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Agent retrieved successfully
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  data:
                    $ref: '#/definitions/agent.Config'
                type: object
        "400":
          description: Invalid agent ID
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  error:
                    $ref: '#/definitions/router.ErrorInfo'
                type: object
        "404":
          description: Agent not found
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  error:
                    $ref: '#/definitions/router.ErrorInfo'
                type: object
        "500":
          description: Internal server error
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  error:
                    $ref: '#/definitions/router.ErrorInfo'
                type: object
      summary: Get agent by ID
      tags:
        - agents
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication (Optional - controlled by SERVER_AUTH_ENABLED environment
      variable. When enabled, all API endpoints require valid authentication. In development/testing,
      authentication may be disabled.)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...